Add credentials validation function
